change to float 64

+++ perceval/utils/format.py
@@ -46,8 +46,8 @@ def simple_float(alpha, precision=1e-6, nsimplify=True, fracmax=63, multiplier=1
     if nsimplify:
         for r in range(1, fracmax):
             for multiplier2 in [sp.S(1), sp.pi, sp.sqrt(2), sp.sqrt(3), sp.sqrt(5), sp.sqrt(6)]:
-                v = np.float128(alpha/multiplier2*r)
-                round_v = float(np.float128(v).round())
+                v = np.float64(alpha/multiplier2*r)
+                round_v = float(np.float64(v).round())
                 if abs(float(v)-round_v) < precision:
                     simple = sign*sp.Rational(round_v, r)*multiplier*multiplier2
                     return simple, str(simple)
@@ -66,7 +66,7 @@ def simple_float(alpha, precision=1e-6, nsimplify=True, fracmax=63, multiplier=1
             alpha = alpha / 10
             mult10 -= 1
 
-    alpha = float(np.float128(alpha/precision).round()) * precision
+    alpha = float(np.float64(alpha/precision).round()) * precision
     simple_str = str(sp.S(alpha))
     if simple_str.find(".") != -1:
         for i in range(len(simple_str)-1, 0, -1):
